requirements.txt
# Core Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# LangGraph & LangChain
langgraph>=0.0.26
langchain>=0.1.6
langchain-community>=0.0.19
langchain-core>=0.1.23

# Vector Database & Embeddings
faiss-cpu>=1.7.4
sentence-transformers>=2.3.1

# PDF & Document Processing
PyPDF2>=3.0.1
python-docx>=1.1.0
pdfminer.six>=20221105
easyocr>=1.7.1
Pillow>=10.2.0

# Database
sqlalchemy>=2.0.25

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.4.0

# API & HTTP
httpx>=0.26.0
python-multipart>=0.0.7
pydantic>=2.5.3
pydantic-settings>=2.1.0

# Visualization
plotly>=5.18.0

# Utilities
python-dotenv>=1.0.0
aiofiles>=23.2.1
tqdm>=4.66.1
tenacity>=8.2.3
